requirements.txt
alembic==1.0.10
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
Mako==1.0.12
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
psycopg2-binary==2.8.3
psycopg2==2.8.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
scipy==1.2.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
